requirements.txt
pystray>=0.19.0
pillow>=9.0.0
pyyaml>=6.0
cryptography>=3.4.0
pyautogui>=0.9.50
psutil>=5.8.0